test(sdp-api): split payments route tests into domain files
